Finish crud routing
